VIREO-103: Managers may update all field labels and sticky notes (the student side)

+++ app/controllers/submit/FileUpload.java
@@ -1,11 +1,22 @@
 package controllers.submit;
 
 import java.io.File;
+import java.io.IOException;
+import java.io.StringReader;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
+import org.tdl.vireo.constant.FieldConfig;
+
+import static org.tdl.vireo.constant.AppConfig.SUBMIT_UPLOAD_FILES_STICKIES;
+import static org.tdl.vireo.constant.FieldConfig.*;
 import org.tdl.vireo.model.Attachment;
+import org.tdl.vireo.model.AttachmentType;
 import org.tdl.vireo.model.RoleType;
 import org.tdl.vireo.model.Submission;
+
+import au.com.bytecode.opencsv.CSVReader;
 
 import controllers.Security;
 import controllers.Student;
@@ -35,28 +46,33 @@ public class FileUpload extends AbstractSubmitStep {
 		// Locate the submission that this upload will be attached to
 		Submission sub = getSubmission();
 		
-		// If the replace manuscript button is pressed - then delete the manuscript 
-		if (params.get("replacePrimary") != null) {
-			Attachment primaryDoc = sub.getPrimaryDocument();   
-			if (primaryDoc != null) {
-				primaryDoc.delete();
-				sub.save();
+		// Upload or replace the primary document.
+		if (isFieldEnabled(PRIMARY_ATTACHMENT)) {
+			if (params.get("replacePrimary") != null) {
+				Attachment primaryDoc = sub.getPrimaryDocument();   
+				if (primaryDoc != null) {
+					primaryDoc.delete();
+					sub.save();
+				}
 			}
+			
+			if(params.get("primaryDocument",File.class) != null)
+				Student.uploadPrimaryDocument(sub);
 		}
 		
-		// Handle the remove supplementary document button 
-		if (params.get("removeSupplementary") != null) {
-			Student.removeSupplementary(sub);           	            	
+		// Manage supplemental files
+		if (isFieldEnabled(SUPPLEMENTARY_ATTACHMENT)) {
+			if (params.get("removeSupplementary") != null)
+				Student.removeSupplementary(sub);           	            	
+			if(params.get("supplementaryDocument",File.class) != null)
+				Student.uploadSupplementary(sub);
 		}
 		
-		if(params.get("primaryDocument",File.class) != null)
-			Student.uploadPrimaryDocument(sub);
-		
-		if(params.get("supplementaryDocument",File.class) != null)
-			Student.uploadSupplementary(sub);
-		
-		if (!"true".equals(flash.get("nextStep")))
-			verify(sub.getPrimaryDocument());
+		// Verify the form if we are submitting or if jumping from the confirm step.
+		if ("fileUpload".equals(params.get("step")) ||
+			"confirm".equals(flash.get("from-step"))) {
+			verify(sub);
+		}
 
 		
 		// 'Save And Continue' button was clicked
@@ -69,27 +85,42 @@ public class FileUpload extends AbstractSubmitStep {
 		Attachment primaryAttachment = sub.getPrimaryDocument();
 		List<Attachment> supplementalAttachments = sub.getSupplementalDocuments();
 
-		renderTemplate("Submit/fileUpload.html",subId, primaryAttachment, supplementalAttachments);
+		List<String> stickies = new ArrayList<String>();
+		String stickiesRaw = settingRepo.getConfigValue(SUBMIT_UPLOAD_FILES_STICKIES);
+		if (stickiesRaw != null && !"null".equals(stickiesRaw)) {
+			try {
+				CSVReader reader = new CSVReader(new StringReader(stickiesRaw));
+				stickies = Arrays.asList(reader.readNext());
+				reader.close();
+			} catch (IOException ioe) {
+				throw new RuntimeException(ioe);
+			}
+		}
+		
+		renderTemplate("Submit/fileUpload.html",subId, primaryAttachment, supplementalAttachments, stickies);
 	}
 
 	/**
 	 * Verify that the user has supplied a primary document. This will be used
 	 * from both the fileUpload form and the confirmation page.
 	 * 
-	 * @param primaryDocument
-	 *            The primary document to confirm it's existence.
 	 * @return True if the primary document exists, otherwise false.
 	 */
-	public static boolean verify(Attachment primaryDocument) {
+	public static boolean verify(Submission sub) {
 		
-		if (primaryDocument == null || primaryDocument.getId() == null) {
+		int numberOfErrorsBefore = validation.errors().size();
+
+		if (isFieldRequired(PRIMARY_ATTACHMENT) && sub.getPrimaryDocument() == null )
 			validation.addError("primaryDocument", "A manuscript file must be uploaded.");
-			return false;
-		} else {
+
+		if (isFieldRequired(SUPPLEMENTARY_ATTACHMENT) && sub.getAttachmentsByType(AttachmentType.SUPPLEMENTAL).size() == 0)
+			validation.addError("supplementaryDocument", "At least one supplementary file is required.");
+
+		if (numberOfErrorsBefore == validation.errors().size()) 
 			return true;
-		}
-		
+		else
+			return false;
 	}
-	
+
 
 }
